+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- kplay will not render its TUI if the terminal dimensions are below the
+  required minimum
+- Pressing Ctrl+C will immediately exit the TUI, regardless of the current
+    context
+
 ## [v3.0.0] - Sep 21, 2025
 
 ### Added